requirements.txt
streamlit>=1.28.0
requests>=2.31.0
folium>=0.14.0
streamlit-folium>=0.13.0
pandas>=2.0.0
streamlit-autorefresh>=0.3.1
matplotlib>=3.7.0
pgeocode>=0.4.0
python-dotenv>=1.0.0
openai>=1.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.0
pytz>=2023.3 